requirements.txt
openai==0.28.1
openai-whisper==20230918
pvcheetah==1.1.3
pvporcupine==2.2.1
pvrecorder==1.2.1
PyAudio==0.2.13
pyautogen==0.1.3
python-dotenv==1.0.0
pyttsx3==2.90
scipy==1.11.3
sounddevice==0.4.6
soundfile==0.12.1
SpeechRecognition==3.10.0
taskwarrior==0.1.2
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tqdm==4.66.1
